Extract
This still differs from PR drafts in how it handles a missing record and upon success, however leaving one permanent IDB record around probably does make sense.
